Modified page style from 3 columns to 2 columns
This is still a temporary setup, but the page was just too wide.  By
dropping to 2 columns, and rearranging things a bit, the site will work
as normal down to 1280px wide, and will drop to a functional, if ugly,
layout below 1280 and remain functional down to about 800px or so.